requirements/base.txt
colorama==0.4.1
colorthief==0.2.1
python-dateutil==2.8.1
discord.py==1.5.1
pymongo==3.9.0
requests==2.22.0
lxml==4.6.2
